[arcane,std] Déplace les services associés à 'BasicReaderWriter' dans le fichier 'BasicCheckpointService.cc'
