+++ arcane/src/arcane/std/srcs.cmake
@@ -1,6 +1,7 @@
 set(ARCANE_SOURCES
   ArcaneStdRegisterer.cc
   ArcaneStdRegisterer.h
+  BasicCheckpointService.cc
   BasicReaderWriter.cc
   BasicReaderWriterDatabase.cc
   BasicReaderWriterDatabase.h
